🚚 Move views / navigation in its own file
